extract ipc types into types.rs and remove unwraps
Move all IPC type definitions (FrameType, StreamFrame, CommandFrame,
SnapshotRegistry, etc.) from ipc/mod.rs into a dedicated ipc/types.rs
module. Re-export everything from mod.rs so callers are unaffected.

Replace the bare .unwrap() on the embedded schema parse in validate.rs
with a ? that maps the error to a String, matching the function's
existing Result<Value, String> return type.
